chore: correct openui5 imports after upgrade to 1.109
