+++ packages/base/package.json
@@ -35,7 +35,7 @@
   },
   "devDependencies": {
     "@buxlabs/amd-to-es6": "0.16.1",
-    "@openui5/sap.ui.core": "1.101.0",
+    "@openui5/sap.ui.core": "1.109.0",
     "@ui5/webcomponents-tools": "1.9.1",
     "chromedriver": "107.0.3",
     "clean-css": "^5.2.2",
